Erase destroyed shader modules from LayerData.  (#63)
Previously, when DestroyShaderModule was called, the module was removed
but its record remained in the LayerData. Since the hash function uses
the pointer and the size of the shader module to create a key, another
shader module that is allocated in the same address and has the same
size could potentially cause a bug. Because its corresponding key will
be the same as the key that already exist in the LayerData.

Remove a shader module from the LayerData when DestroyShaderModule is
called. This makes sure that we don't keep record of the destroyed
objects.

Add a hook on DestroyShaderModule function in the compile_time_layer and
runtime_layer.

Fixes: https://github.com/googlestadia/performance-layers/issues/51