Refactor check-layers.yml for better resource utilization. (#65)
The current CI is checking for every combination of config, compiler,
and generator. The checks contain redundancy without introducing more
coverage. To reduce the redundant checks, only a subset of the
combinations should be executed.

The building for both ninja and Makefiles is redundant, so we mainly
build for ninja, but have a single check for Makefiles.

Keep both gcc and clang to exploit their different optimizations and
diagnostic features, and keep the checks independent of
compiler-specific behaviors. No need to compile with both debug and
release flags for both compilers.

Set the default config, compiler, and generator in the job's matrix. The
include filed expands the matrix by adding `Makefiles, gcc, Release` and
`Ninja, gcc, Debug` configurations.