Update dockerfile to install Vulkan SDK.
Performance layers are not tested on a Vulkan application. Here, the
Vulkan application is `vkcube`. To test the layers on a docker, the
Vulkan application must be executed in the headless mode. `xvfb` allows
us to run in the headless mode by performing all the graphical operaions
in memory instead of showing them on a display.