Update dockerfile to install Vulkan SDK. (#66)
Performance layers are not tested on a Vulkan application. Here, the
Vulkan application is `vkcube`. To test the layers on a docker, the
Vulkan application must be executed in headless mode. `xvfb` allows us
to run in headless mode by performing all the graphical operations in
memory instead of showing them on a display.