Test if enabled layers are populating their log files. (#68)
Setting the environment variables and running an application does not
necessarily guarantee that the layers are enabled. Because the program
can still run even if the environment variables are set incorrectly.
There are many ways that allow us to check whether the layers are
enabled or not.
We can run `vulkaninfo` and look at the `layers` subsection. We can also
use `grep <name-of-the-layer>` to check whether or not they exist in
`vulkaninfo`.
Moreover, when a layer is enabled, it produces logs and stores them in
its log file which can be specified by setting its corresponding
environment variable. By inspecting these log files we can determine the
execution of the layer.

Since, in the next steps, we'd like to make sure that the logs are
consistent with each other (and the input arguments), we've decided to
use the second approach (inspecting the log files). The log files
contain a header row followed by the data row(s). Here, we only examine
the generation of data rather than the content of data. Hence, we only
need to make sure that the log files contain a data row (has at least
two rows).