Add checks for the content of the compile time log.
Previously, we checked whether the log files were populated or not. This
does not necessarily gaurantee that they contain the correct data. Since
we know the pattern in which the data was generated, we can check the
content of the log file and see if it is as expected.
There are many tools (e.g, FileCheck, OutputCheck, etc) that allow us to
perform the pattern matching. We can also write a python script that reads
the log file and matches it against a given pattern. We have picked
FileCheck because it is well maintained and easy to use. It receives two
files and uses one to verify the other. Here, the `check_compile_time_log.txt`
contains the patterns we expect to see in the compile_time_log.csv. The
patterns can be literal (line 3) or contain a regular expression (line 4).
If the patterns don't match, FileCheck will exit with a non-zero value.