Add checks for the content of the compile time log. (#70)
Previously, to verify the layers are working on an application, we only
checked whether the log files were populated or not. This does not
necessarily guarantee that the logs contain the correct data. Since we
know the pattern in which the data was generated, we can check the
content of the log file and see if it is as expected. There are many
tools that allow us to perform the pattern matching (e.g, LLVM
FileCheck, OutputCheck, etc). We can also write a python script that
reads the log file and matches it against a given pattern.
We have selected FileCheck because it is well maintained and easy to
use. Moreover, we don't need to write the code that reads the input
files, stores the data, etc.
FileCheck receives two files and uses one to verify the other. Here, the
`check_compile_time_log.txt` contains the patterns we expect to see in
the compile_time_log.csv. The patterns can be literal (line 3) or
contain a regular expression (line 4). If the patterns don't match,
FileCheck will exit with a non-zero value.