Add checks to the event log file contents. (#77)
In addition to layer-specified log files, all the layers write their
logs into a shared log file (called event log). Like other log files,
this file can be specified by setting an environment variable
(`VK_RUNTIME_LOG`).

`Check_event_log.txt` contains the patterns against which the event log
is tested. It checks the format and count of rows. Also, it uses string
substitution to make sure the hash values are consistent across
different lines. `SHADER1` and `SHADER2` are variables that are set when
the shader modules are created. They are used in the subsequent lines to
verify that the logged hash values remain the same.

Since the layers can be executed in any order (based on their order in
`VK_INSTANCE_LAYERS`) `CHECK-DAG` is used to look for the patterns
regardless of their precedence. For instance, if `frame_time_layer_init`
is moved before `compiler_time_layer_init`, the check will pass. The
order of `CHECK-DAG`s does not matter until we have another type of
check. For instance, the `CHECK` at line 10 acts as a barrier for the
`CHECK-DAG`s before it.