Implement EventLog struct for creating logs.
Currently, the logging mecahnism is unstructured. Various log functions
with different interfaces are used to generate the logs. All these
functions recieve series of inputs, combine them in a string with csv
format, and write them in the standard output or a file.

EventLog unifies the interface for creating logs. It provides these two
functionalities:
1) Using the same interface for creating all the logs to replace the
current ad-hoc event logging functions.
2) Allowing log generation for different formats other than csv.

Each EventLog has a name and a set of attributes. Attributes are
key-value pairs. The key is a string representing what the attribute is
(e.g, duration, timestamp, etc) and the value is the data corresponding
to the key.

The current log functions use only a limited set of types; string, int64,
and vector. Hence, right now, `Attribute` has three implementations
covering the mentioned types.

This PR also contains the implementation for EventLog, an even-to-csv
generator, and simple unit tests to check if they work as expected.