Implement Event class for creating logs.  (#79)
Currently, the logging mechanism is unstructured. Various log functions
with different interfaces are used to generate the logs. All these
functions receive a series of inputs, combine them in a string with csv
format, and write them in the standard output or a file.
`The`Event` class unifies the interface for creating logs. It provides
these two functionalities:
1) Using the same interface for creating all the logs to replace the
current ad-hoc event logging functions.
2) Allowing log generation for different formats other than csv.

Each `Event` has a name and a set of attributes. Attributes contain the
value and its name. The name is a string telling what the attribute is
(e.g, duration, timestamp, etc).

The current log functions use only a limited set of types; string,
int64, and vector. Hence, right now, `Attribute` has three
implementations covering the mentioned types.

This PR also contains the implementation for Event, an event-to-csv
generator, and simple unit tests to check if they work as expected.

Issue: #75