Add implementation for `CSVLogger`.
Some events produced by the layers should be written in CSV. `CSVLogger`
is an implementation of `EventLogger` that takes events and generates
logs in CSV format. It receives the CSV header in its constructor and
writes it to the output when `StartLog()` is called. On `AddEvent()`, it
converts the incoming event to a CSV string(a line of text) and writes it
to the output.