Rafactor memory usage layer to use CSVLogger.
Add `MemoryUsageEvent` to memory usage layer.

This PR implements the `MemoryUsageEvent` class. It is an event that
contains the allocation information such as current and peak allocated
bytes. The data should be logged to both private and the common files.
`MemoryUsageEvent` is used in both `QueuePresentKHR()` and `DestroyDevice()`
instead of the current `LogUsage()` call.

`MemoryUsageLayerData` allocates and deallocates The `CSVLogger`. The
logger accepts events through the new `LogEvent` method. `LogEvent` takes a
`MemoryUsageEvent` as input and logs it to the memory layer's log file.