Refactor the code to use chrono time instead of absl.
One of the future directions of perfomancelayers is to have dynamic
unloading to provide better support for the applications. Currently, the
`absl` libraries we're using don't support dynamic unloading. Hence, we
should not use `absl` in the new features and refactor the current
codebase to use alternative libraries.

This PR refactors the performancelayers to use `chrono` time instead of
`absl` time.

`chrono::steady_clock`, `chrono::system_clock`,
`chrono::high_resolution_clock`, and `absl::Now()` brief explanation:
1. `steady_clock`: A monotonic clock. The current value of steady_clock
does not matter, but the guarantee that is strictly increasing is useful
for calculating the difference between two samples.
2. `system_clock`: Is the wall-clock. Used for time/day related stuff.
3. `high_resolution_clock`: An alias for one of the above.
4. `absl::Now()`: Constructs an absl::Time from the system clock.

Previously, `absl::Now()` was used for the timestamp and measuring
duration. This PR separates these two by using `system_clock::now()` for
the timestamp and `steady_clock::now()` for the duration.