Refactor the code to use chrono time instead of absl. (#94)
One of the future directions of perfomancelayers is to have dynamic
unloading to provide better support for the applications. Currently, the
`absl` libraries we're using don't support dynamic unloading. Hence, we
should not use `absl` in the new features and refactor the current
codebase to use alternative libraries.

This PR refactors the performancelayers to use `chrono` time instead of
`absl` time.