Add support for duration attribute.
By using an `Int64Attr` as a single number that represents time-related
values (duration/timestamp), we would loose the time unit information.

The `DurationAttr` is a class that contains duration information. It is a
wrapper around a `chrono::steady_clock::duration` variable. The duration
can be converted into time units. The logger decides what time unit it
should be converted into.