Add support for timestamp attribute.
By using an Int64Attr as a single number that represents time-related
values (duration/timestamp), we would loose the time unit information.

The `TimestampAttr` is an attribute that contains duration information.
It is a wrapper around a `chrono::system_clock::time_point` variable.
This variable has nanoseconds precision and the loggers are responsible
for converting it to the desired time unit.