Enable `warning-as-error` and fix the errors.
Some of the current uncaptured warnings cause failures in the imports
becuase the imports consider some of the  warnings as errors.

This PR sets the `warning-as-error` flag in the `cmake` (except for
`comment` and `return-type` warnings`) and fixes some
the existing warnings such as missing initializations, and unhandled
switch-case statements.