Enable `warning-as-error` and fix the errors. (#102)
Some of the current unhandled warnings cause failures in the imports
because the imports consider them as errors.

This PR enables warnings in the `cmake` and sets the `warning-as-error`
flag in the CI (except for `comment` and `return-type` warnings`) so
that the CI can catch the warnings that would break the imports. It also
fixes some existing warnings such as missing initializations, unhandled
switch-case statements, and unclosed files.