Add `LogOutput` class.
This class provides an abstraction over the output. All the subclasses
must implement `LogLine()` method which receives a string and writes it
to the output.

This PR also contains 2 implementaions of `LogOutput`class. `FileOutput`
handles writing into a file and standard output. `StringOutput` writes to
a string and is used for testing.