Fix `vkcube` hanging problem with runtime layer enabled.
On ubuntu 22.04, `vkcube` didn't finish when `runtime layer` was enable.
Running `valgrind` showed that `cmd_buff` was accessed after it got
freed by `vkcube`. The `cube.c` calls `vkFreeCommandBuffers()` which
frees the command buffer. Runtime layer tried using the command buffer
without knowing it was freed.

In this PR, runtime layer overrides the `vkFreeCommandBuffer` to capture
the list of command buffers about to be freed. The `RemoveQueries()`
method takes these command buffers and erases the queries that have them.