Fix `vkcube` hanging problem with runtime layer enabled. (#120)
On ubuntu 22.04, `vkcube` didn't finish when `runtime layer` was
enabled. Running `valgrind` showed that `cmd_buff` was accessed after it
got freed by `vkcube`.
```
==2139036== Invalid read of size 8
==2139036==    at 0x15316C85: performancelayers::RuntimeLayerData::LogAndRemoveQueryPools() (in /usr/local/google/home/miladhakimi/performance-layers/build/run/lib/libVkLayer_stadia_pipeline_runtime.so)
==2139036==    by 0x15311D37: (anonymous namespace)::RuntimeLayer_DeviceWaitIdle(VkDevice_T*) (in /usr/local/google/home/miladhakimi/performance-layers/build/run/lib/libVkLayer_stadia_pipeline_runtime.so)
==2139036==    by 0x404B1C: demo_cleanup (cube.c:2392)
==2139036==    by 0x404B1C: main (cube.c:4327)
==2139036==  Address 0x8322320 is 0 bytes inside a block of size 5,680 free'd
==2139036==    at 0x484617B: free (vg_replace_malloc.c:872)
==2139036==    by 0x14C323A3: ??? (in /usr/lib/x86_64-linux-gnu/libvulkan_lvp.so)
==2139036==    by 0x404B01: demo_cleanup (cube.c:2386)
==2139036==    by 0x404B01: main (cube.c:4327)
```
The `cube.c` calls `vkFreeCommandBuffers()` which frees the command
buffer. The runtime layer then uses the command buffers without knowing
it was freed. Here is a part of the code of `cube.c` where it frees the
command buffer.
```
for (i = 0; i < demo->swapchainImageCount; i++) {
    vkFreeCommandBuffers(demo->device, demo->cmd_pool, 1, &demo->swapchain_image_resources[i].cmd);
    vkDestroyBuffer(demo->device, demo->swapchain_image_resources[i].uniform_buffer, NULL);
}
```

To fix that, we intercept `vkFreeCommandBuffer()` in the runtime layer
and get the list of command buffers about to be freed. The
`RemoveQueries()` method takes these command buffers and erases the
queries that have them.

Fixes: #71