Refactor `StringOutput` to take the ownership of the log strings.
In the `AddEvent()` method, the logger generates a log string on the
stack and passes it as a `string_view` to the `LogOutput`. If the logger
writes to a file, there is no need to access the stack-allocated log
string anymore as it's persisted inside the file. `StringOutput` works
differently. It keeps the logs in a `vector<string_view>` and reads them
after we're done with logging. Hence, it needs to have access to the
logs after we return from `AddEvent()` method.
To fix this, we copy the logs in the `StringOutput` and keep a
`vector<string>` instead of `string_view`.