Refactor `StringOutput` to keep a copy of log strings. (#123)
In the `AddEvent()` method, the logger generates a log string on the
stack and passes it as a `string_view` to the `LogOutput`. If the logger
uses `FileOutput`, using `string_view` is fine because there is no need
to access the stack-allocated log string after returning from
`AddEvent()` as it is persisted to the file. However, `StringOutput`
works differently. It keeps the logs in a `vector<string_view>` and
reads them after we're done with logging. Hence, it needs to access the
logs after we return from the `AddEvent()`.
To fix this, we copy the logs in the `StringOutput` and keep a
`vector<string>` instead of `string_view`.