Fix formatting problem in float to string conversion.
`ostringstream` mixes the formatting of the floats. To represent floats,
instead of relying on the stream to handle the float to string conversion,
we use `std::to_chars` with `fixed` format.

The size of the string representing the number is unknown and varies from
one input to the other. Hence, the input buffer given to `std::to_chars`
is dynamically allocated and resized to fit in the number.

The size of the buffer might become more than the size of the characters
written to it. Although all the rest of characters are '\0', writing it
to a `ostringstream` coppies the whole buffer because it's using the
size of the string to perform the copy. By calling the `.data()`, the
function copies the characters up until the first '\0' and does not copy
the rest.