Ignore trace events in CSV logger (#152)
Currently, trace events are ignored by the CSV logger, but the delimiter
appending logic does not take that into account, resulting in a
delimiter being appended to every line with an empty value after. That
causes CSV files not to match the declared header format anymore.

The invalid output looks like this:
```
Frame Time (ns),Benchmark State
16614739,1,
```

This PR makes it so that the CSV logger filters out events that it can't
log (only trace events for now). Both unit and FileCheck tests are
updated to check for this scenario.

The CommonLogger still outputs such events, but given the fact that it's
somewhat more structured (outputs attributes in the format
`attribute_name:attribute_value`), I feel it's ok if it outputs an
attribute with an empty value as a parser has enough information to skip
over it.