Fix the slack event bug. (#155)
A slack event shows the time between creation of a shader and its first
use. `ShaderModuleSlack` is used to track these values using
`creation_end_time` and `first_use_time` variables. Since the timestamp
for an event shows the end of its creation, the `creation_end_time`
should be set by the `CreateShaderEvent`'s timestamp. `first_use_time`
should be set by the timestamp of the start of the pipeline.

Fixes: #150